# services/settlement-service/src/main/java/com/sportsbetting/settlementservice/dto/SettleEventRequest.java
package com.sportsbetting.settlementservice.dto;

import jakarta.validation.constraints.NotBlank;

public record SettleEventRequest(
        @NotBlank String eventId,
        @NotBlank String winningSelection
) {
}