# services/settlement-service/src/main/java/com/sportsbetting/settlementservice/dto/SettleEventResponse.java
package com.sportsbetting.settlementservice.dto;

import java.math.BigDecimal;

public record SettleEventResponse(
        String eventId,
        String winningSelection,
        int winners,
        int losers,
        BigDecimal totalPayout,
        BigDecimal globalExposure
) {
}